+++ frontend/components/EditBookmarkModal.vue
@@ -1,4 +1,5 @@
 <script setup lang="ts">
+import { useAuthStore } from '@/stores/AuthStore';
 import type { Bookmark, Tag, Folder } from '@/types';
 
 const props = defineProps({
@@ -18,38 +19,66 @@ const props = defineProps({
 
 const emit = defineEmits<{
   (e: "close", isVisible: boolean): void;
+  (e: "updated", isUpdated: boolean): void;
 }>();
 
+const authStore = useAuthStore();
+const config = useRuntimeConfig();
+
 const editableBookmark: Ref<Bookmark> = ref(props.bookmark);
-const selectedFolder: Ref<number> = ref(0);
+const selectedFolderID: Ref<number> = ref(0);
 
 if (props.bookmark.folder?.id) {
-  selectedFolder.value = props.bookmark.folder.id;
+  selectedFolderID.value = props.bookmark.folder.id;
+} else {
+  // This is to prevent weird 400 errors when patching with `null` value.
+  editableBookmark.value.folder = undefined;
 }
 
 function closeModal() {
   emit("close", false);
 }
 
-function onClickSaveChanges() {
+async function onClickSaveChanges() {
   /*
     Send updated bookmark data to API using "PATCH" method (partial update).
   */
 
-  if (selectedFolder.value) {
-    /*
-      If a folder was selected, save it's `id` in bookmark object. We don't care
-      about user id and folder title because backend will refer to folder id (which is unique)
-      to get it from the DB and assign to the bookmark.
-    */
-    editableBookmark.value.folder = {
-      id: selectedFolder.value,
-      user: 0,
-      title: "",
-    }
+  if (selectedFolderID.value) {
+    // Find folder object with selected id
+    const selectedFolder = props.allFolders.filter((folder) => folder.id === selectedFolderID.value)[0];
+    editableBookmark.value.folder = selectedFolder;
+  } else {
+    editableBookmark.value.folder = undefined;
   }
 
-  console.log("editableBookmark = ", editableBookmark.value);
+  const formData = {
+    url: editableBookmark.value.url,
+    title: editableBookmark.value.title,
+    description: editableBookmark.value.description,
+    image_url: editableBookmark.value.image_url,
+    folder: editableBookmark.value.folder,
+    is_favorite: editableBookmark.value.is_favorite,
+    is_read: editableBookmark.value.is_read,
+    is_archived: editableBookmark.value.is_archived,
+  }
+
+  const { data: bookmarkData, error: bookmarkUpdateError } = await useFetch(() => `${config.public.apiBase}/api/v1/bookmarks/update/${props.bookmark.id}/`, {
+    method: "PATCH",
+    headers: [
+      ["Authorization", "Token " + authStore.token,],
+    ],
+    body: formData,
+  });
+
+  if (bookmarkUpdateError.value) {
+    console.error("Error updating bookmark data: " + bookmarkUpdateError.value?.message);
+    alert("Something went wrong. Please try again!");
+    return;
+  }
+
+  emit("updated", true);
+  closeModal();
 }
 </script>
 
@@ -69,6 +98,13 @@ function onClickSaveChanges() {
             </div>
 
             <div class="field">
+              <label class="label">Bookmark URL</label>
+              <div class="control">
+                <input class="input" type="text" placeholder="Cover image URL" v-model="editableBookmark.url">
+              </div>
+            </div>
+
+            <div class="field">
               <label class="label">Image URL</label>
               <div class="control">
                 <input class="input" type="text" placeholder="Cover image URL" v-model="editableBookmark.image_url">
@@ -83,9 +119,10 @@ function onClickSaveChanges() {
               </div>
             </div>
           </div>
+
           <div class="column is-4">
             <!-- Cover image -->
-            <figure class="image is-3by2 mt-6" v-if="editableBookmark.image_url">
+            <figure class="image is-16by9 mt-6" v-if="editableBookmark.image_url">
               <img :src="editableBookmark.image_url">
             </figure>
           </div>
@@ -99,11 +136,11 @@ function onClickSaveChanges() {
               <div class="control">
                 <div class="select">
 
-                  <select v-model="selectedFolder">
+                  <select v-model="selectedFolderID">
                     <option value="0">
                       None
                     </option>
-                    <option v-for="folder in allFolders" v-bind:value="folder.id">
+                    <option v-for="folder in allFolders" v-bind:value="folder.id" :key="`folder-id-${folder.id}`">
                       {{ folder.title }}
                     </option>
                   </select>
@@ -139,7 +176,8 @@ function onClickSaveChanges() {
               <div class="control">
 
                 <BulmaTagList v-if="editableBookmark.tags.length">
-                  <BulmaTag v-for="tag in editableBookmark.tags" :tag="tag" :hasDeleteButton="true">
+                  <BulmaTag v-for="tag in editableBookmark.tags" :tag="tag" :hasDeleteButton="true"
+                    :key="`assigned-tag-${tag.id}`">
                     {{ tag.title }}
                   </BulmaTag>
                 </BulmaTagList>
@@ -155,7 +193,8 @@ function onClickSaveChanges() {
               <div class="control">
 
                 <BulmaTagList v-if="allTags.length">
-                  <BulmaTag v-for="tag in allTags" :tag="tag" :hasDeleteButton="false" :hasCounter="true">
+                  <BulmaTag v-for="tag in allTags" :tag="tag" :hasDeleteButton="false" :hasCounter="true"
+                    :key="`all-tags-${tag.id}`">
                     {{ tag.title }}
                   </BulmaTag>
                 </BulmaTagList>